+++ src/main/java/com/team/winners/thirdcycleproject/controllers/ProfileController.java
@@ -25,7 +25,7 @@ public class ProfileController {
     }
 
     @DeleteMapping (value = "delete/{id}")
-    private ResponseEntity<Boolean> deleteProfile (@PathVariable ("id") Long id){
+    private ResponseEntity<Boolean> deleteProfile (@PathVariable ("id") Integer id){
         return ResponseEntity.ok(profileService.deleteById(id));
     }
